Integrate kpath.py into lattice_models/tools.py
